requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
openai
python-dotenv
streamlit
